Move indicator management into ExploreModel
