+++ charts/__tests__/ExploreModel.test.ts
@@ -1,5 +1,8 @@
 import { ExploreModel } from "../ExploreModel"
 import { ChartType } from "../ChartType"
+import { RootStore } from "charts/Store"
+
+const store = new RootStore()
 
 describe(ExploreModel, () => {
     let model: ExploreModel
@@ -7,7 +10,7 @@ describe(ExploreModel, () => {
     describe("when you give it a query string", () => {
         beforeAll(() => {
             const queryStr = "indicator=488&time=1960..2000&type=WorldMap"
-            model = new ExploreModel()
+            model = new ExploreModel(store)
             model.populateFromQueryStr(queryStr)
         })
 
@@ -26,7 +29,7 @@ describe(ExploreModel, () => {
 
     describe("when you set an area type", () => {
         beforeAll(() => {
-            model = new ExploreModel()
+            model = new ExploreModel(store)
             model.chartType = ChartType.StackedArea
         })
 
@@ -49,7 +52,7 @@ describe(ExploreModel, () => {
 
     describe("when you set a map type", () => {
         beforeAll(() => {
-            model = new ExploreModel()
+            model = new ExploreModel(store)
             model.chartType = ExploreModel.WorldMap
         })
